[inferbo] Revise eval_locs for array blocks
Summary:
It renames `eval_locs` to `eval_arr` and we use it for getting array block values the given input expressions are pointing to.  For example, when given a program variable `x` as an input, `eval_arr` returns array blocks that `x` is pointing to, on the other hand, `eval` returns an abstract location of `x`.

Depends on D7471891

Reviewed By: mbouaziz

Differential Revision: D7471915

fbshipit-source-id: b994944