[sqlite] use pname BLOB as the sqlite key, include readable pname for debug
Summary:
We were using the "filename" as the key because it's (kinda) unique *and* human
readable, but with the `infer explore --procedures` interface we don't really
need the human readable part anymore, so we can just use the OCaml marshalling
of the pname as the key. The human-readable version (sans unique-fying hash) is
now another column in the table, used to match procedure names in
`--procedures-filter`.

Reviewed By: sblackshear

Differential Revision: D7639158

fbshipit-source-id: e714605