[absint] fix handling of exceptions
Summary:
The abstract interpreter tried to handle exceptional control-flow by propagating the *pre* of a block that threw an exception rather than the *post*.
This was a half-measure that isn't correct when an exception-throwing instruction isn't in the middle of a block.
The handling of exceptions wasn't actually used anywhere and was leading to further hacks in `ProcCfg`, so let's get rid of it.

Reviewed By: mbouaziz, jvillard

Differential Revision: D7843872

fbshipit-source-id: 2a4a815