[starvation] deduplication
Summary:
There can be multiple reports per line, especially when calling in a method which has itself multiple reports.

When reporting at the callsite, report only the issue with the highest severity (picked manually per type of event).

Deadlocks are not de-duplicated, as they are relatively rare and the info in mupltiple reports may be important.

Reviewed By: mbouaziz

Differential Revision: D8160940

fbshipit-source-id: ea6a5c0