[tests] record error bucket in expected output
Summary:
It's useful to test that the bucket a given error is classified as doesn't
change over time without notice.

This records the bucket for *all* the tests, even though some never produce a
bucket. This is to be on the safe size instead of risking to forget adding the
bucket information when the test changes, or when copy/pasting from a test that
doesn't have buckets to one that does.

The implementation is pretty crude: it greps the beginning of the qualifier
string for a `[bucket]`.

Reviewed By: mbouaziz

Differential Revision: D8236393

fbshipit-source-id: b3b1eb9