[starvation] report using split events
Summary:
Use the component of the abstract state `events` to report. This set contains reachability facts about blocking calls and lock acquisitions.

The other component, `order`, contained pairs of a reachable event `e'` from an event option with the semantics that if the option is `None` then we have an element that now goes into `events`, and if the option is `Some e` then the element represents a lock acquired and a trace *from* `e` to `e'`

Now, `order` can be simplified to proper pairs of events, without the option.

Reviewed By: jvillard

Differential Revision: D8227665

fbshipit-source-id: e6f4dac