[backend] remove source file from exe_env
Summary:
The execution environment is really just a cache. It happens to point to a
particular source file which is where the analysis was started from, but that
is not relevant, and in fact is confusing because it suggests that it is
somewhat tied to that file. In reality, exe_env caches information about any
procedure and source file encountered by the analysis.

This will make it easier to make further changes but I think it also brings a
bit more clarity to the code.

Reviewed By: jeremydubreil

Differential Revision: D8513735

fbshipit-source-id: f4b38ce