[backend] do not let caches grow too big
Summary:
This changes the behaviour back to before the "fork once then pipe new tasks
for each source files" era: before that we would fork once per source file,
which had the effect of emptying the caches for every source file. Without this
the caches can grow unchecked.

This is probably suboptimal but should at least be the same behaviour as before.

Reviewed By: jeremydubreil

Differential Revision: D8615124

fbshipit-source-id: 69fc101