[cli] add the ability to rerun the analysis on some files with different options
Summary:
When `--reanalyze` is passed, mark the summaries of procedures matching
`--procedures-filter` as needing to be analysed before running the analysis.

This allows one to, for instance, re-run the analysis in debug mode on only
some files or procedures. However, this won't work for the Java Buck
integration since the summaries are hidden away in buck-out.

Reviewed By: mbouaziz

Differential Revision: D8783668

fbshipit-source-id: 9032d83