[Inferbo] Add relational domain
Summary:
It adds relational domains to Inferbo: octagon of Apron and polyhedra of Elina.

- Each `Mem` domain value includes one relational value containing relations among *symbols*.  The relational values are modified by the `Prune` and `Store` commands.
- Each abstract value includes three *symbols*, which represent integer value, array offset, and array size of an abstract value.

The relational domain is deactivated by default, so this diff should not make any differences in CI.

Use `--bo-relational-domain {oct, poly}` for the activation, though Inferbo with the relational domains does not work at this point because some modifications of Apron and Elina we made has not been applied to their opam repositories yet.

Reviewed By: mbouaziz, jvillard

Differential Revision: D8478542

fbshipit-source-id: 510ff53