[taskbar] less flicker and garbage
Summary:
`ANSITerminal` flushes after almost every operation. As a result, there is a
lot of flicker in the task bar. Going through `Format` instead reduces the
flicker.

Also erase end-of-lines when outputting logging stuff to the console. This makes sure the output is not garbled with stuff from the task bar.

Reviewed By: ngorogiannis

Differential Revision: D8931294

fbshipit-source-id: e89c87f84