[infer] fix the internal confusion between the issue kind and the severity
Summary:
The internal concept of "kind" should in fact be named "severity" to match the convention used by many other tools, whereas the internal concept of "severity", i.e "HIGH", "MEDIUM" and "LOW" was never used and in any case redundant with the concept of "info", "warning", "error".

This diff maps both the "kind" and "severity" fields to value of the form "advice", "info", "warning", and "error" to be able to progressively migrate the code using the "kind" field.

Reviewed By: mbouaziz, jvillard

Differential Revision: D9187978

fbshipit-source-id: 447d89f51