[caml] do not use `let _ =`
Summary:
Use `ignore` instead, as this will warn if the argument is an arrow type,
unlike `let _ = ...`. This makes the code more future-proof: if an argument is
added to a function called in `let _ = f x` then the compiler will complain
instead of silently turning a value into a partial evaluation.

Also got rid of particularly irksome `let _ = <stuff returning unit> in` where I could.

Reviewed By: mbouaziz

Differential Revision: D9217176

fbshipit-source-id: 3be463405