[epilogues] do not rely on `at_exit`
Summary:
For some unexplained reason, some of the functions registered in the Epilogues would sometimes be executed several times. I could not figure out why.

This diff fixes that, but also has more explainable benefits:
- Do not run epilogues registered in the parent in the children. Previously it
  would do so, but probably only if the children registered some epilogue given
  that `at_exit` must be called again once on the child (but the value of the ref
  in `Pervasives` would not have been reset).
- Unified behaviour for early and late epilogues given that we now handle both of these directly

We already have all the control needed to run epilogues when needed: we know
when infer exits, and we know when children processes exit.

Reviewed By: mbouaziz

Differential Revision: D9752046

fbshipit-source-id: 13af40081