[eradicate] no longer report Field Not Nullable warnings on the Fragment onDestroy method
Summary: It is common on Android code to recycle the `View` object by nullifying them in the `onDestroy()` or `onDestroyView()` methods. In this case, the outer `Fragment` object structure is preserve while the inner `View` object are set to null for the garbage collect to release the memory. However, if the fields are only set to `null` in the `onDestroy*()` methods, those fields cannot be `null` during the active lifecycle of the `Fragment`, so it is not necessary to annotate those fields with `Nullable`.

Reviewed By: mbouaziz

Differential Revision: D10024458

fbshipit-source-id: b05e538d9