[absint] unify html debug for HIL and SIL
Summary:
HIL wanted to do its own HTML printing, causing code duplication and hacks to
avoid double opening/closing files. Instead, pass a hook to print SIL
instructions or not.

This also makes the debug HTML be printed even in case of raised exceptions,
which is invaluable to debug crashes or even just reports in the case of
checkers that can raise `Stop_analysis` (pulse only for now).

This also print intermediate abstract states between instructions instead of
only at the start and end of nodes, for moar debugging.

Reviewed By: mbouaziz

Differential Revision: D12857425

fbshipit-source-id: 4ee6c88d6